requirements.txt
anyascii==0.3.2
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-filter==22.1
django-jalali-date==1.0.2
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.4
install==1.3.5
jdatetime==4.1.1
l18n==2021.3
mysqlclient==2.1.1
oauthlib==3.2.2
openpyxl==3.1.2
passlib==1.7.4
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
urllib3==2.0.2
wagtail==5.0
webencodings==0.5.1
Willow==1.5